+++ src/constants/index.js
@@ -32,12 +32,6 @@ export const UNDER_MAINTENANCE = false;
 
 export const GA_ID = "GTM-MXXQHG8";
 
-export const DEVICE_TYPE_DETAILS = {
-  computer: "Default screen 1366px width, 768px height",
-  tablet: "Default screen (vertical) 810px height, 1080px width",
-  phone: "Default screen 375px width, 812px height",
-};
-
 export const ROUTES = {
   INTAKE_FORM: "/self-service/wizard",
   HOME_PAGE: "/self-service",
@@ -583,34 +577,6 @@ export const ACTIONS = {
 export const AUTO_SAVE_FORM = "AUTO_SAVE_FORM";
 
 export const CACHED_CHALLENGE_ID = "CACHED_CHALLENGE_ID";
-
-// TODO: these values are being used by both legacy and new web design.
-// New web design should probably use the ones in webDesignConfigs
-export const CHALLENGE_FIELD_VALUES = {
-  trackId: "5fa04185-041f-49a6-bfd1-fe82533cd6c8",
-  typeId: "927abff4-7af9-4145-8ba1-577c16e64e2e",
-  timelineTemplateId: "918f6a3e-1a63-4680-8b5e-deb95b1411e7",
-};
-
-// TODO: this timeline is being used by both legacy and new web design.
-// New web design should probably use the one in webDesignConfigs
-export const DEFAULT_TIMELINE = [
-  {
-    // Screening
-    phaseId: "2d7d3d85-0b29-4989-b3b4-be7f2b1d0aa6",
-    duration: 14400,
-  },
-  {
-    // Review
-    phaseId: "aa5a3f78-79e0-4bf7-93ff-b11e8f5b398b",
-    duration: 144000,
-  },
-  {
-    // Approval
-    phaseId: "ad985cff-ad3e-44de-b54e-3992505ba0ae",
-    duration: 14400,
-  },
-];
 
 export const CHALLENGE_STATUS = {
   ACTIVE: "Active",
